Separate logging functions into logging module
